requirements.txt
#tkinter
Pillow
moviepy